BAU Scale out if there are no desired-instances
We have had a couple of occasions where the bastion was scaled in, but not scaled out again.
The code here tries to determine if it needs to scale out by looking if there are zero running instances.
A better check here is to look if the _desired_ instances is zero - otherwise an instance that is slow to terminate could stop the autorecycle from scaling out again.